refactored code , fixed navbar
